Rename SectionInputDataFrame -> SectionArrayInput and move it to new file
