+++ test/entities/test_arrays.py
@@ -6,7 +6,7 @@ from numpy.testing import assert_allclose
 from pandas.testing import assert_frame_equal
 from pandera.typing import DataFrame
 
-from mechaphlowers.entities.arrays import SectionArray, SectionInputDataFrame
+from mechaphlowers.entities.arrays import SectionArray, SectionArrayInput
 
 
 @pytest.fixture
@@ -24,19 +24,19 @@ def section_array_input_data() -> dict[str, list]:
 
 @pytest.fixture
 def section_array(section_array_input_data: dict[str, list]) -> SectionArray:
-    df = DataFrame[SectionInputDataFrame](section_array_input_data)
+    df = DataFrame[SectionArrayInput](section_array_input_data)
     return SectionArray(data=df, sagging_parameter=2_000, sagging_temperature=15)
 
 
 def test_create_section_array__with_floats(section_array_input_data: dict) -> None:
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
     section = SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
 
     assert_frame_equal(input_df, section._data, check_dtype=False, rtol=1e-07)
 
 
 def test_create_section_array__only_ints() -> None:
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(
+    input_df: DataFrame[SectionArrayInput] = DataFrame(
         {
             "name": ["support 1", "2", "three", "support 4"],
             "suspension": [False, True, True, False],
@@ -56,7 +56,7 @@ def test_create_section_array__span_length_for_last_support(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["span_length"][-1] = 300
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -64,7 +64,7 @@ def test_create_section_array__span_length_for_last_support(
 
 def test_create_section_array__missing_name(section_array_input_data: dict) -> None:
     del section_array_input_data["name"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -74,7 +74,7 @@ def test_create_section_array__missing_suspension(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["suspension"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -84,7 +84,7 @@ def test_create_section_array__missing_conductor_attachment_altitude(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["conductor_attachment_altitude"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -94,7 +94,7 @@ def test_create_section_array__missing_crossarm_length(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["crossarm_length"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -104,7 +104,7 @@ def test_create_section_array__missing_line_angle(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["line_angle"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -114,7 +114,7 @@ def test_create_section_array__missing_insulator_length(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["insulator_length"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -124,7 +124,7 @@ def test_create_section_array__missing_span_length(
     section_array_input_data: dict,
 ) -> None:
     del section_array_input_data["span_length"]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -135,7 +135,7 @@ def test_create_section_array__extra_column(
 ) -> None:
     section_array_input_data["extra column"] = [0] * 4
 
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     section = SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
 
@@ -146,7 +146,7 @@ def test_create_section_array__wrong_name_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["name"] = [1, 2, 3, 4]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -156,7 +156,7 @@ def test_create_section_array__wrong_suspension_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["suspension"] = [1, 2, 3, 4]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -166,7 +166,7 @@ def test_create_section_array__wrong_conductor_attachment_altitude_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["conductor_attachment_altitude"] = ["1,2"] * 4
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -176,7 +176,7 @@ def test_create_section_array__wrong_crossarm_length_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["crossarm_length"] = ["1,2"] * 4
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -186,7 +186,7 @@ def test_create_section_array__wrong_line_angle_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["line_angle"] = ["1,2"] * 4
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -196,7 +196,7 @@ def test_create_section_array__wrong_insulator_length_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["insulator_length"] = ["1,2"] * 4
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -206,7 +206,7 @@ def test_create_section_array__wrong_span_length_type(
     section_array_input_data: dict,
 ) -> None:
     section_array_input_data["span_length"] = ["1,2"] * 4
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
@@ -217,14 +217,14 @@ def test_create_section_array__insulator_length_for_tension_support(
 ) -> None:
     section_array_input_data["suspension"] = [False, False, True, False]
     section_array_input_data["insulator_length"] = [0.5, 0.5, 0.5, 0.5]
-    input_df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    input_df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     with pytest.raises(pa.errors.SchemaErrors):
         SectionArray(input_df, sagging_parameter=2_000, sagging_temperature=15)
 
 
 def test_compute_elevation_difference(section_array_input_data: dict) -> None:
-    df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
 
     section_array = SectionArray(df, sagging_parameter=2_000, sagging_temperature=15)
 
@@ -234,7 +234,7 @@ def test_compute_elevation_difference(section_array_input_data: dict) -> None:
 
 
 def test_section_array__data(section_array_input_data: dict) -> None:
-    df: DataFrame[SectionInputDataFrame] = DataFrame(section_array_input_data)
+    df: DataFrame[SectionArrayInput] = DataFrame(section_array_input_data)
     section_array = SectionArray(
         data=df, sagging_parameter=2_000, sagging_temperature=15
     )
